schema: give private objects a durable identity
## Why

Private object storage still relied on serial-style identifiers and left size and digest invariants to application code.

## What changed

- use identity-backed bigint IDs and text columns for unbounded keys and media types
- preserve key uniqueness and binary payload storage
- enforce nonnegative sizes and exact hexadecimal SHA-256 digests
- add an additive migration that realigns the sequence and keeps existing rows valid
- regenerate the database model

## Testing

- full unit suite
- race suite
- static analysis
- sqlc generation and module tidy diff
- disposable database migration rehearsal

+++ server/internal/platform/database/postgres/models.go
@@ -125,7 +125,7 @@ type PassengerReview struct {
 }
 
 type PrivateObject struct {
-	ID             int32              `db:"id"`
+	ID             int64              `db:"id"`
 	StorageKey     string             `db:"storage_key"`
 	Content        []byte             `db:"content"`
 	ContentType    string             `db:"content_type"`
